services/registry-api/pyproject.toml
[project]
name = "registry-api"
version = "0.1.0"
description = "RYE Registry API - Server-side validation and signing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "supabase>=2.3.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "rye-os>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["registry_api"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
